Calculate the c-value of the difference in PCCs produced by two groups

Calculate the c-value of the difference in PCCs produced by two groups

Usage

compare_groups(m, group1, group2)

Arguments

m

an object of class "opafit" produced by a call to opa().

group1

a character string which matches a group level passed to opa().

group2

a character string which matches a group level passed to opa().

Value

an object of class "opaGroupComparison".

Examples

dat <- data.frame(group = c("a", "b", "a", "b"),
                  t1 = c(9, 4, 8, 10),
                  t2 = c(8, 8, 12, 10),
                  t3 = c(8, 5, 10, 11))
dat$group <- factor(dat$group, levels = c("a", "b"))
opamod <- opa(dat[,2:4], 1:3, group = dat$group)
compare_groups(opamod, "a", "b")
#> $group1
#> [1] "a"
#> 
#> $group2
#> [1] "b"
#> 
#> $group1_pcc
#>        a 
#> 33.33333 
#> 
#> $group2_pcc
#>        b 
#> 66.66667 
#> 
#> $pcc_diff
#> [1] 33.33333
#> 
#> $cval
#> [1] 0.395
#> 
#> $pcc_diff_dist
#>    [1]  1.666667e+01 -5.000000e+01 -1.666667e+01 -3.333333e+01  3.333333e+01
#>    [6] -5.000000e+01 -3.333333e+01  1.666667e+01 -3.333333e+01 -3.333333e+01
#>   [11] -3.333333e+01  5.000000e+01 -3.333333e+01  1.666667e+01  5.000000e+01
#>   [16]  3.333333e+01  0.000000e+00  0.000000e+00  0.000000e+00  7.105427e-15
#>   [21]  1.666667e+01  1.666667e+01  0.000000e+00 -3.333333e+01 -1.666667e+01
#>   [26]  1.666667e+01  1.666667e+01  7.105427e-15  3.333333e+01 -5.000000e+01
#>   [31]  3.333333e+01 -3.333333e+01  1.666667e+01  0.000000e+00  0.000000e+00
#>   [36]  0.000000e+00 -1.666667e+01  3.333333e+01 -1.666667e+01  1.666667e+01
#>   [41]  1.666667e+01 -3.333333e+01  5.000000e+01  0.000000e+00 -3.333333e+01
#>   [46]  0.000000e+00  5.000000e+01 -3.333333e+01  0.000000e+00  1.666667e+01
#>   [51]  5.000000e+01  3.333333e+01  3.333333e+01  0.000000e+00  1.666667e+01
#>   [56]  1.666667e+01 -3.333333e+01  0.000000e+00 -1.666667e+01  0.000000e+00
#>   [61]  1.666667e+01  1.666667e+01  3.333333e+01  1.666667e+01  5.000000e+01
#>   [66] -3.333333e+01  1.666667e+01 -1.666667e+01 -5.000000e+01  1.666667e+01
#>   [71]  3.333333e+01 -1.666667e+01  1.666667e+01  1.666667e+01 -1.666667e+01
#>   [76] -6.666667e+01  1.666667e+01 -3.333333e+01 -3.333333e+01  1.666667e+01
#>   [81]  5.000000e+01 -3.333333e+01  0.000000e+00  0.000000e+00  1.666667e+01
#>   [86]  3.333333e+01  3.333333e+01 -6.666667e+01  3.333333e+01  5.000000e+01
#>   [91] -1.666667e+01  1.666667e+01  1.666667e+01  0.000000e+00  1.666667e+01
#>   [96]  0.000000e+00 -3.333333e+01 -1.666667e+01  6.666667e+01  7.105427e-15
#>  [101] -1.666667e+01  1.666667e+01  1.666667e+01 -1.666667e+01 -8.333333e+01
#>  [106]  0.000000e+00  5.000000e+01 -5.000000e+01  0.000000e+00 -3.333333e+01
#>  [111] -3.333333e+01  0.000000e+00  6.666667e+01 -3.333333e+01  1.666667e+01
#>  [116] -3.333333e+01 -1.666667e+01  1.666667e+01  1.666667e+01  0.000000e+00
#>  [121]  0.000000e+00  0.000000e+00  3.333333e+01  5.000000e+01  3.333333e+01
#>  [126]  1.666667e+01 -1.666667e+01  3.333333e+01 -3.333333e+01  8.333333e+01
#>  [131] -1.666667e+01 -6.666667e+01  3.333333e+01 -1.666667e+01 -7.105427e-15
#>  [136]  3.333333e+01  0.000000e+00 -5.000000e+01  5.000000e+01 -1.666667e+01
#>  [141]  5.000000e+01  0.000000e+00  0.000000e+00  0.000000e+00  1.666667e+01
#>  [146] -1.666667e+01  0.000000e+00 -5.000000e+01  0.000000e+00 -5.000000e+01
#>  [151]  0.000000e+00  1.666667e+01  1.666667e+01  0.000000e+00 -3.333333e+01
#>  [156] -3.333333e+01 -1.666667e+01  3.333333e+01 -6.666667e+01  3.333333e+01
#>  [161] -1.666667e+01  5.000000e+01  3.333333e+01  0.000000e+00 -1.666667e+01
#>  [166]  0.000000e+00  1.666667e+01 -1.666667e+01  1.666667e+01  1.666667e+01
#>  [171]  1.666667e+01  0.000000e+00  0.000000e+00  5.000000e+01  0.000000e+00
#>  [176]  1.666667e+01 -5.000000e+01  1.666667e+01  3.333333e+01  3.333333e+01
#>  [181]  0.000000e+00 -3.333333e+01  1.666667e+01 -3.333333e+01 -1.666667e+01
#>  [186] -1.666667e+01 -1.666667e+01  3.333333e+01  1.666667e+01  3.333333e+01
#>  [191]  3.333333e+01  1.666667e+01  0.000000e+00  1.666667e+01 -1.666667e+01
#>  [196]  1.666667e+01  3.333333e+01 -3.333333e+01  1.666667e+01  3.333333e+01
#>  [201] -3.333333e+01  0.000000e+00  3.333333e+01  0.000000e+00  0.000000e+00
#>  [206]  5.000000e+01  1.666667e+01  0.000000e+00  3.333333e+01 -1.666667e+01
#>  [211]  5.000000e+01 -5.000000e+01 -1.666667e+01 -1.666667e+01  3.333333e+01
#>  [216]  3.333333e+01  1.666667e+01  5.000000e+01  1.666667e+01  1.666667e+01
#>  [221]  6.666667e+01 -1.666667e+01  1.666667e+01  1.666667e+01  5.000000e+01
#>  [226]  1.666667e+01 -1.666667e+01 -1.666667e+01  3.333333e+01 -7.105427e-15
#>  [231]  1.666667e+01  0.000000e+00 -6.666667e+01 -1.666667e+01  0.000000e+00
#>  [236]  0.000000e+00  1.666667e+01  1.666667e+01  0.000000e+00  0.000000e+00
#>  [241]  1.666667e+01  1.666667e+01  0.000000e+00  3.333333e+01  0.000000e+00
#>  [246]  3.333333e+01  0.000000e+00 -5.000000e+01 -6.666667e+01  0.000000e+00
#>  [251]  1.666667e+01 -1.666667e+01  3.333333e+01  0.000000e+00  1.666667e+01
#>  [256] -5.000000e+01  0.000000e+00  6.666667e+01 -3.333333e+01  5.000000e+01
#>  [261]  0.000000e+00 -1.666667e+01 -6.666667e+01  0.000000e+00  1.666667e+01
#>  [266]  0.000000e+00  0.000000e+00  0.000000e+00  1.666667e+01 -1.666667e+01
#>  [271] -1.666667e+01  0.000000e+00  1.666667e+01  5.000000e+01 -1.666667e+01
#>  [276]  0.000000e+00  0.000000e+00  3.333333e+01 -1.666667e+01  3.333333e+01
#>  [281]  3.333333e+01 -3.333333e+01  3.333333e+01 -5.000000e+01  7.105427e-15
#>  [286]  0.000000e+00 -1.666667e+01  3.333333e+01 -1.666667e+01 -3.333333e+01
#>  [291] -5.000000e+01  3.333333e+01 -1.666667e+01  5.000000e+01 -1.666667e+01
#>  [296]  0.000000e+00 -1.666667e+01 -1.666667e+01 -1.666667e+01  0.000000e+00
#>  [301]  0.000000e+00  0.000000e+00  0.000000e+00 -3.333333e+01  6.666667e+01
#>  [306]  0.000000e+00  3.333333e+01  3.333333e+01  1.666667e+01 -5.000000e+01
#>  [311]  0.000000e+00  1.666667e+01 -7.105427e-15  1.666667e+01 -5.000000e+01
#>  [316]  1.666667e+01  0.000000e+00 -1.666667e+01  1.666667e+01  1.666667e+01
#>  [321]  3.333333e+01  0.000000e+00 -3.333333e+01 -3.333333e+01 -3.333333e+01
#>  [326] -5.000000e+01  3.333333e+01 -3.333333e+01 -3.333333e+01 -1.666667e+01
#>  [331]  0.000000e+00  1.666667e+01 -3.333333e+01 -1.666667e+01 -3.333333e+01
#>  [336]  1.666667e+01  1.666667e+01 -3.333333e+01 -3.333333e+01  1.666667e+01
#>  [341] -5.000000e+01  0.000000e+00  1.666667e+01 -1.666667e+01  1.666667e+01
#>  [346]  0.000000e+00 -3.333333e+01  1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [351]  5.000000e+01 -5.000000e+01  3.333333e+01  3.333333e+01  0.000000e+00
#>  [356]  3.333333e+01  1.666667e+01  1.666667e+01 -3.333333e+01 -1.666667e+01
#>  [361]  1.666667e+01 -5.000000e+01 -3.333333e+01 -3.333333e+01  3.333333e+01
#>  [366]  0.000000e+00 -3.333333e+01  3.333333e+01 -1.666667e+01  1.666667e+01
#>  [371]  1.666667e+01  1.666667e+01  0.000000e+00 -7.105427e-15  1.666667e+01
#>  [376] -1.666667e+01 -3.333333e+01  3.333333e+01  1.666667e+01 -1.666667e+01
#>  [381]  3.333333e+01 -5.000000e+01  3.333333e+01  0.000000e+00  0.000000e+00
#>  [386] -3.333333e+01 -5.000000e+01 -3.333333e+01 -5.000000e+01  8.333333e+01
#>  [391] -6.666667e+01  3.333333e+01 -1.666667e+01  3.333333e+01  6.666667e+01
#>  [396]  1.666667e+01 -3.333333e+01  0.000000e+00 -3.333333e+01  1.666667e+01
#>  [401]  1.666667e+01  5.000000e+01  6.666667e+01 -3.333333e+01  1.666667e+01
#>  [406]  3.333333e+01 -3.333333e+01  3.333333e+01 -5.000000e+01  0.000000e+00
#>  [411]  3.333333e+01 -7.105427e-15  1.666667e+01 -5.000000e+01 -1.666667e+01
#>  [416]  1.666667e+01 -1.666667e+01 -3.333333e+01  0.000000e+00 -3.333333e+01
#>  [421]  3.333333e+01  0.000000e+00  0.000000e+00  0.000000e+00  1.666667e+01
#>  [426]  1.666667e+01 -1.666667e+01  5.000000e+01  5.000000e+01  1.666667e+01
#>  [431] -3.333333e+01  0.000000e+00  1.666667e+01  1.666667e+01  1.666667e+01
#>  [436]  0.000000e+00  0.000000e+00 -1.666667e+01  6.666667e+01  0.000000e+00
#>  [441]  1.666667e+01 -1.666667e+01  1.666667e+01 -1.666667e+01  6.666667e+01
#>  [446]  3.333333e+01  3.333333e+01  0.000000e+00  0.000000e+00 -5.000000e+01
#>  [451] -3.333333e+01  1.666667e+01  1.666667e+01 -1.666667e+01 -3.333333e+01
#>  [456]  0.000000e+00 -3.333333e+01 -1.666667e+01  3.333333e+01 -1.666667e+01
#>  [461]  5.000000e+01  3.333333e+01  1.666667e+01  1.666667e+01  3.333333e+01
#>  [466]  3.333333e+01  1.666667e+01 -1.666667e+01 -3.333333e+01  0.000000e+00
#>  [471] -1.666667e+01 -3.333333e+01  1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [476]  3.333333e+01 -1.666667e+01  3.333333e+01 -1.666667e+01  5.000000e+01
#>  [481]  3.333333e+01 -3.333333e+01  1.666667e+01 -5.000000e+01 -3.333333e+01
#>  [486]  3.333333e+01  1.666667e+01 -1.666667e+01  1.666667e+01  1.666667e+01
#>  [491]  1.666667e+01  3.333333e+01  0.000000e+00  3.333333e+01  3.333333e+01
#>  [496] -1.666667e+01  5.000000e+01  3.333333e+01  3.333333e+01 -1.666667e+01
#>  [501]  3.333333e+01 -7.105427e-15  5.000000e+01  0.000000e+00 -3.333333e+01
#>  [506]  0.000000e+00 -1.666667e+01  0.000000e+00 -1.666667e+01 -3.333333e+01
#>  [511]  3.333333e+01  0.000000e+00  1.666667e+01 -3.333333e+01  5.000000e+01
#>  [516] -1.666667e+01 -3.333333e+01 -3.333333e+01 -7.105427e-15 -3.333333e+01
#>  [521]  1.666667e+01 -1.666667e+01  1.666667e+01 -5.000000e+01 -5.000000e+01
#>  [526]  1.666667e+01 -6.666667e+01  6.666667e+01  1.666667e+01 -1.666667e+01
#>  [531] -7.105427e-15  3.333333e+01  1.666667e+01  3.333333e+01 -3.333333e+01
#>  [536]  1.666667e+01  3.333333e+01  3.333333e+01 -1.666667e+01  3.333333e+01
#>  [541] -1.666667e+01  5.000000e+01  1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [546]  1.666667e+01  6.666667e+01  0.000000e+00 -7.105427e-15  0.000000e+00
#>  [551]  0.000000e+00  3.333333e+01 -5.000000e+01 -5.000000e+01 -3.333333e+01
#>  [556] -3.333333e+01 -1.666667e+01  0.000000e+00 -3.333333e+01  3.333333e+01
#>  [561]  5.000000e+01  1.666667e+01  1.666667e+01 -1.666667e+01  0.000000e+00
#>  [566] -1.666667e+01 -1.666667e+01  3.333333e+01  1.666667e+01  1.666667e+01
#>  [571]  0.000000e+00  0.000000e+00 -1.666667e+01  0.000000e+00  1.666667e+01
#>  [576]  0.000000e+00 -6.666667e+01  0.000000e+00 -3.333333e+01  5.000000e+01
#>  [581] -3.333333e+01  0.000000e+00  0.000000e+00 -3.333333e+01  1.666667e+01
#>  [586]  1.666667e+01 -6.666667e+01  1.666667e+01 -3.333333e+01  3.333333e+01
#>  [591]  0.000000e+00 -3.333333e+01  1.666667e+01  0.000000e+00  1.666667e+01
#>  [596] -1.666667e+01  3.333333e+01 -1.666667e+01 -3.333333e+01  3.333333e+01
#>  [601]  1.666667e+01 -3.333333e+01  3.333333e+01 -1.666667e+01  0.000000e+00
#>  [606]  1.666667e+01  5.000000e+01 -1.666667e+01 -8.333333e+01  3.333333e+01
#>  [611]  7.105427e-15  3.333333e+01 -3.333333e+01  1.666667e+01  0.000000e+00
#>  [616]  5.000000e+01  5.000000e+01  5.000000e+01  0.000000e+00  7.105427e-15
#>  [621] -6.666667e+01  0.000000e+00 -1.666667e+01  1.666667e+01 -3.333333e+01
#>  [626]  6.666667e+01  0.000000e+00  5.000000e+01  1.666667e+01  1.666667e+01
#>  [631] -3.333333e+01 -3.333333e+01  0.000000e+00 -1.666667e+01  1.666667e+01
#>  [636]  5.000000e+01 -1.666667e+01  0.000000e+00 -3.333333e+01 -1.666667e+01
#>  [641]  0.000000e+00  1.666667e+01 -1.666667e+01 -5.000000e+01 -5.000000e+01
#>  [646]  5.000000e+01  1.666667e+01 -1.666667e+01  0.000000e+00 -5.000000e+01
#>  [651]  5.000000e+01 -6.666667e+01  1.666667e+01  0.000000e+00  1.666667e+01
#>  [656] -1.666667e+01  3.333333e+01 -1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [661]  0.000000e+00  0.000000e+00  1.666667e+01  0.000000e+00  0.000000e+00
#>  [666]  3.333333e+01 -3.333333e+01 -1.666667e+01  0.000000e+00  0.000000e+00
#>  [671] -3.333333e+01  0.000000e+00  0.000000e+00 -1.666667e+01 -1.666667e+01
#>  [676] -3.333333e+01  1.666667e+01  0.000000e+00  1.666667e+01  1.666667e+01
#>  [681]  7.105427e-15  1.666667e+01 -5.000000e+01  0.000000e+00 -3.333333e+01
#>  [686]  0.000000e+00 -5.000000e+01 -1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [691]  0.000000e+00  0.000000e+00 -1.666667e+01  1.666667e+01 -1.666667e+01
#>  [696]  3.333333e+01  0.000000e+00  1.666667e+01 -1.666667e+01  7.105427e-15
#>  [701]  1.666667e+01 -1.666667e+01  1.666667e+01  0.000000e+00 -1.666667e+01
#>  [706]  0.000000e+00 -1.666667e+01  0.000000e+00  0.000000e+00 -1.666667e+01
#>  [711]  3.333333e+01 -5.000000e+01  7.105427e-15  3.333333e+01  1.666667e+01
#>  [716]  3.333333e+01  3.333333e+01  6.666667e+01  0.000000e+00  1.666667e+01
#>  [721] -5.000000e+01  5.000000e+01 -5.000000e+01  0.000000e+00  0.000000e+00
#>  [726]  3.333333e+01 -1.666667e+01  0.000000e+00 -1.666667e+01 -7.105427e-15
#>  [731]  1.666667e+01 -1.666667e+01 -8.333333e+01  3.333333e+01  0.000000e+00
#>  [736]  0.000000e+00  0.000000e+00  0.000000e+00  1.666667e+01  3.333333e+01
#>  [741] -3.333333e+01 -1.666667e+01  1.666667e+01 -1.666667e+01 -3.333333e+01
#>  [746] -1.666667e+01  0.000000e+00  1.666667e+01 -1.666667e+01  0.000000e+00
#>  [751] -1.666667e+01  1.666667e+01 -5.000000e+01  1.666667e+01 -1.666667e+01
#>  [756]  1.666667e+01  1.666667e+01 -1.666667e+01  0.000000e+00  1.666667e+01
#>  [761] -1.666667e+01 -3.333333e+01  1.666667e+01  3.333333e+01  1.666667e+01
#>  [766]  3.333333e+01 -3.333333e+01  1.666667e+01  1.666667e+01  0.000000e+00
#>  [771] -1.666667e+01 -5.000000e+01 -1.666667e+01 -5.000000e+01 -1.666667e+01
#>  [776]  6.666667e+01 -1.666667e+01 -6.666667e+01 -1.666667e+01 -7.105427e-15
#>  [781] -1.666667e+01  1.666667e+01  5.000000e+01 -7.105427e-15 -5.000000e+01
#>  [786]  1.666667e+01  0.000000e+00  1.666667e+01 -3.333333e+01  3.333333e+01
#>  [791] -1.666667e+01  1.666667e+01  5.000000e+01 -1.666667e+01  3.333333e+01
#>  [796]  1.666667e+01 -1.666667e+01 -3.333333e+01 -1.666667e+01  5.000000e+01
#>  [801]  1.666667e+01  1.666667e+01 -1.666667e+01 -5.000000e+01  0.000000e+00
#>  [806]  0.000000e+00  0.000000e+00 -1.666667e+01 -3.333333e+01 -1.666667e+01
#>  [811]  0.000000e+00  6.666667e+01 -1.666667e+01 -3.333333e+01  0.000000e+00
#>  [816] -1.666667e+01 -5.000000e+01  3.333333e+01 -1.666667e+01  0.000000e+00
#>  [821]  1.666667e+01 -3.333333e+01  3.333333e+01  1.666667e+01  5.000000e+01
#>  [826]  7.105427e-15  3.333333e+01 -3.333333e+01 -5.000000e+01  0.000000e+00
#>  [831]  0.000000e+00  3.333333e+01  1.666667e+01 -1.666667e+01  1.666667e+01
#>  [836]  1.666667e+01 -1.666667e+01  0.000000e+00  3.333333e+01  0.000000e+00
#>  [841] -1.666667e+01  1.666667e+01  0.000000e+00  0.000000e+00  1.666667e+01
#>  [846]  1.666667e+01 -1.666667e+01  0.000000e+00  3.333333e+01  1.666667e+01
#>  [851] -5.000000e+01  1.666667e+01  1.666667e+01 -1.666667e+01  8.333333e+01
#>  [856] -1.666667e+01 -1.666667e+01 -1.666667e+01 -1.666667e+01  0.000000e+00
#>  [861]  1.666667e+01  0.000000e+00  1.666667e+01 -3.333333e+01  3.333333e+01
#>  [866]  6.666667e+01 -3.333333e+01 -7.105427e-15  0.000000e+00 -1.666667e+01
#>  [871]  3.333333e+01 -1.666667e+01  5.000000e+01 -1.666667e+01 -1.666667e+01
#>  [876]  0.000000e+00  0.000000e+00 -3.333333e+01  0.000000e+00 -1.666667e+01
#>  [881] -1.666667e+01 -1.666667e+01 -1.666667e+01  1.666667e+01  3.333333e+01
#>  [886]  0.000000e+00 -3.333333e+01 -3.333333e+01 -5.000000e+01  1.666667e+01
#>  [891]  5.000000e+01  1.666667e+01  0.000000e+00 -1.666667e+01 -5.000000e+01
#>  [896]  1.666667e+01  1.666667e+01  6.666667e+01  5.000000e+01 -1.666667e+01
#>  [901] -1.666667e+01  5.000000e+01  5.000000e+01  0.000000e+00 -5.000000e+01
#>  [906]  1.666667e+01  1.666667e+01 -3.333333e+01 -5.000000e+01 -1.666667e+01
#>  [911] -1.666667e+01  1.666667e+01  0.000000e+00 -5.000000e+01 -5.000000e+01
#>  [916]  1.666667e+01  3.333333e+01  5.000000e+01 -1.666667e+01  3.333333e+01
#>  [921] -7.105427e-15  1.666667e+01 -1.666667e+01 -1.666667e+01  1.666667e+01
#>  [926]  0.000000e+00 -5.000000e+01  0.000000e+00 -3.333333e+01  6.666667e+01
#>  [931] -1.666667e+01  1.666667e+01 -1.666667e+01 -3.333333e+01  0.000000e+00
#>  [936]  1.666667e+01  3.333333e+01  0.000000e+00  0.000000e+00  0.000000e+00
#>  [941] -1.666667e+01  0.000000e+00 -5.000000e+01  1.666667e+01  1.666667e+01
#>  [946]  1.666667e+01 -7.105427e-15  0.000000e+00 -5.000000e+01  1.666667e+01
#>  [951]  5.000000e+01  1.666667e+01 -3.333333e+01  0.000000e+00 -3.333333e+01
#>  [956] -3.333333e+01  5.000000e+01  3.333333e+01  1.666667e+01  5.000000e+01
#>  [961]  0.000000e+00 -6.666667e+01  1.666667e+01 -3.333333e+01  1.666667e+01
#>  [966] -1.666667e+01  1.666667e+01 -3.333333e+01 -3.333333e+01 -3.333333e+01
#>  [971]  0.000000e+00  0.000000e+00  1.666667e+01 -1.666667e+01 -1.666667e+01
#>  [976]  0.000000e+00 -3.333333e+01 -1.666667e+01 -3.333333e+01 -6.666667e+01
#>  [981]  3.333333e+01 -1.666667e+01  5.000000e+01 -1.666667e+01  5.000000e+01
#>  [986]  6.666667e+01 -3.333333e+01 -3.333333e+01  7.105427e-15 -1.666667e+01
#>  [991]  1.666667e+01  0.000000e+00  1.666667e+01  3.333333e+01  5.000000e+01
#>  [996]  6.666667e+01  5.000000e+01  1.666667e+01 -5.000000e+01  1.666667e+01
#> 
#> attr(,"class")
#> [1] "opaGroupComparison"